Moved end2end models into 'models' subdir
