+++ depthcharge/__init__.py
@@ -1,8 +1,8 @@
 """Initialize the depthcharge package"""
 from .version import _get_version
-from . import embed
+from . import models
+from . import components
 from . import data
-from . import denovo
 from . import utils
 
 __version__ = _get_version()
