--- depthcharge/denovo/__init__.py
@@ -1,5 +1,0 @@
-"""The de novo sequencing model"""
-from .model import Spec2Pep
-from .dataloaders import DeNovoDataModule
-
-# from .train import train
