+++ depthcharge/models/denovo/__init__.py
@@ -1,0 +1,3 @@
+"""The de novo sequencing model"""
+from .model import Spec2Pep
+from .dataloaders import Spec2PepDataModule
