+++ depthcharge/models/embed/__init__.py
@@ -1,4 +1,4 @@
 """Initialize the embed module"""
-from .model import SiameseSpectrumEncoder
+from .model import PairedSpectrumEncoder
 from .dataloaders import PairedSpectrumDataModule
 from . import performance
